Move things to new class BlockCopier
