+++ src/main/java/com/mraof/minestuck/entry/EntryProcess.java
@@ -26,17 +26,12 @@ import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.chunk.ChunkAccess;
 import net.minecraft.world.level.chunk.LevelChunk;
-import net.minecraft.world.level.chunk.LevelChunkSection;
-import net.minecraft.world.level.levelgen.Heightmap;
 import net.minecraft.world.phys.AABB;
-import net.minecraftforge.registries.ForgeRegistries;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -45,8 +40,6 @@ import java.util.*;
 public class EntryProcess
 {
 	private static final Logger LOGGER = LogManager.getLogger();
-	
-	private static final Set<EntryBlockProcessing> blockProcessors = new HashSet<>();
 	
 	private EntryProcess(ServerPlayer player)
 	{
@@ -58,14 +51,6 @@ public class EntryProcess
 		zDiff = -origin.getZ();
 		
 		creative = player.gameMode.isCreative();
-	}
-	
-	/**
-	 * Not thread-safe. Make sure to only call this on the main thread
-	 */
-	public static void addBlockProcessing(EntryBlockProcessing processing)
-	{
-		blockProcessors.add(processing);
 	}
 	
 	private final int artifactRange = MinestuckConfig.SERVER.artifactRange.get();
@@ -206,7 +191,7 @@ public class EntryProcess
 			
 			if(block.is(Blocks.BEDROCK) || block.is(Blocks.NETHER_PORTAL))
 				block = Blocks.AIR.defaultBlockState();
-			copyBlock(level1, chunk, pos, block, level1.getChunk(pos1), pos1);
+			BlockCopier.copyBlock(level1, chunk, pos, block, level1.getChunk(pos1), pos1);
 		}
 	}
 	
@@ -356,31 +341,6 @@ public class EntryProcess
 		return true;
 	}
 	
-	private static void copyBlockDirect(LevelAccessor levelAccessor, ChunkAccess cSrc, ChunkAccess cDst, int xSrc, int ySrc, int zSrc, int xDst, int yDst, int zDst)
-	{
-		BlockPos dest = new BlockPos(xDst, yDst, zDst);
-		LevelChunkSection blockStorageSrc = getBlockStorage(cSrc, ySrc);
-		LevelChunkSection blockStorageDst = getBlockStorage(cDst, yDst);
-		int y = yDst;
-		xSrc &= 15; ySrc &= 15; zSrc &= 15; xDst &= 15; yDst &= 15; zDst &= 15;
-		
-		boolean isEmpty = blockStorageDst.hasOnlyAir();
-		BlockState state = blockStorageSrc.getBlockState(xSrc, ySrc, zSrc);
-		blockStorageDst.setBlockState(xDst, yDst, zDst, state);
-		if(isEmpty != blockStorageDst.hasOnlyAir())
-			levelAccessor.getChunkSource().getLightEngine().updateSectionStatus(dest, blockStorageDst.hasOnlyAir());    //I assume this adds or removes a light storage section here depending on if it is needed (because a section with just air doesn't have to be regarded)
-		
-		cDst.getOrCreateHeightmapUnprimed(Heightmap.Types.MOTION_BLOCKING).update(xDst, y, zDst, state);
-		cDst.getOrCreateHeightmapUnprimed(Heightmap.Types.MOTION_BLOCKING_NO_LEAVES).update(xDst, y, zDst, state);
-		cDst.getOrCreateHeightmapUnprimed(Heightmap.Types.OCEAN_FLOOR).update(xDst, y, zDst, state);
-		cDst.getOrCreateHeightmapUnprimed(Heightmap.Types.WORLD_SURFACE).update(xDst, y, zDst, state);
-	}
-	
-	private static LevelChunkSection getBlockStorage(ChunkAccess c, int y)
-	{
-		return c.getSection(c.getSectionIndex(y));
-	}
-	
 	/**
 	 * Gives the Y-value of the highest non-air block within artifact range of the coordinates provided in the given world.
 	 */
@@ -407,37 +367,5 @@ public class EntryProcess
 	{
 		GateBlock.placeGate(level, new BlockPos(0, GateHandler.GATE_HEIGHT_1, 0), GateHandler.Type.GATE_1, 0);
 		GateBlock.placeGate(level, new BlockPos(0, GateHandler.GATE_HEIGHT_2, 0), GateHandler.Type.GATE_2, 0);
-	}
-	
-	private static void copyBlock(ServerLevel level, LevelChunk chunkFrom, BlockPos source, BlockState block, ChunkAccess chunkTo, BlockPos dest)
-	{
-		if(chunkTo.getBlockState(dest).is(Blocks.BEDROCK))
-			return;
-		
-		if(block.isAir())
-			level.setBlock(dest, block, 0);
-		else
-			copyBlockDirect(level, chunkFrom, chunkTo, source.getX(), source.getY(), source.getZ(), dest.getX(), dest.getY(), dest.getZ());
-		
-		BlockEntity blockEntity = chunkFrom.getBlockEntity(source, LevelChunk.EntityCreationType.CHECK);
-		BlockEntity newBE = null;
-		if(blockEntity != null)
-		{
-			CompoundTag nbt = blockEntity.saveWithId();
-			nbt.putInt("x", dest.getX());
-			nbt.putInt("y", dest.getY());
-			nbt.putInt("z", dest.getZ());
-			newBE = BlockEntity.loadStatic(dest, block, nbt);
-			if(newBE != null)
-				level.setBlockEntity(newBE);
-			else
-				LOGGER.warn("Unable to create a new block entity {} when teleporting blocks to the medium!", ForgeRegistries.BLOCK_ENTITY_TYPES.getKey(blockEntity.getType()));
-			
-		}
-		
-		for(EntryBlockProcessing processing : blockProcessors)
-		{
-			processing.copyOver((ServerLevel) chunkFrom.getLevel(), source, level, dest, block, blockEntity, newBE);
-		}
 	}
 }
